refactor(courses): standardize module component names and modal behavior
Rename module-related components and files for clarity and
consistency, such as `CourseModulePanel` to `ModuleListPanel`.

Move course-related hooks like `useTreeConnectors` from
`course-editor` to the `courses` directory for broader use.

Enhance `CardWithModal` and `ButtonWithModal` components to
dynamically apply `fullScreen`, `radius`, and `transitionProps`
based on screen size, improving mobile responsiveness and UX.
Adjust `CourseCreationModal` sizing to `lg`.
