+++ frontend/src/features/courses/useSubmissionDetails.ts
